Fix an issue where gifs that failed to convert wouldn't display at all.
Removed webRequest permissions, giving up on properly blocking convertable gifs -
there appears to be a display bug in Chrome where changing the URL of a gif causes
it to load, but stop animating. This is pretty obnoxious and not worth the hassle
or the permissions spam, so I've removed it.